melhoramos a identação do projeto e padronizamos os nomes de componentes e pastas em inglês
