+++ src/components/Logo/index.jsx
@@ -1,7 +1,7 @@
 import { Link } from "react-router-dom";
 import { Image } from "@chakra-ui/react";
 
-function Logo({ size, imagem }){
+function Logo({ size, imagem }) {
     return (
         <Link to="/" display="block">
             <Image src={imagem} lt="Logo zozoz" width={size} />
