+++ src/main.jsx
@@ -4,9 +4,9 @@ import App from "./App.jsx";
 import { ChakraProvider } from "@chakra-ui/react";
 
 ReactDOM.createRoot(document.getElementById("root")).render(
-  <React.StrictMode>
-    <ChakraProvider>
-      <App />
-    </ChakraProvider>
-  </React.StrictMode>
+	<React.StrictMode>
+		<ChakraProvider>
+			<App />
+		</ChakraProvider>
+	</React.StrictMode>
 );
